chain/requirements.txt
web3==6.20.1
py-solc-x==2.0.3
python-dotenv==1.0.1